chore: fix all clippy warnings
- Remove unused imports (auto-fixed)
- Replace manual strip_prefix patterns in doc.rs with strip_prefix()
- Fix never_loop in parser.rs (loop with immediate break → if let .next())
- Convert needless range loops to iterator-based forms
- Simplify redundant boolean expression in complete.rs
- Add type alias MatchEntry for complex type in parser.rs
- Fix doc comment overindentation across multiple files
- Convert outer /// doc to //! inner doc in task_runner.rs
- Suppress too_many_arguments on rg_find_references (intentional API)

+++ src/indexer/scan.rs
@@ -703,7 +703,7 @@ impl Indexer {
                             .ok()
                             .and_then(|v| v.parse::<u128>().ok())
                             .unwrap_or(1000);
-                    if took as u128 > threshold {
+                    if took > threshold {
                         log::warn!(
                             "Slow parse: {} took {} ms",
                             item.path.display(),
